mysite/requirements.txt
asgiref==3.7.2
autopep8==2.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.15.1
idna==3.7
pycodestyle==2.11.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.1
